refactor: move set cookies func to utils package
