+++ .env.example
@@ -10,8 +10,8 @@ CORS_ALLOWED_ORIGINS=http://localhost:3000,http://localhost:5173,https://$UI_URL
 
 PORT=8080
 
-SSO_GOOGLE_KEY=<google-key>
-SSO_GOOGLE_SECRET=<google-secret>
+GOOGLE_SSO_KEY=<google-key>
+GOOGLE_SSO_SECRET=<google-secret>
 SESSION_SECRET=<session-secret>
 
 BASE_PATH=
