+++ pkg/user/util.go
@@ -3,10 +3,6 @@ package user
 import (
 	"context"
 	"fmt"
-	"net/http"
-
-	"github.com/dhis2-sre/im-manager/pkg/token"
-	"github.com/gin-gonic/gin"
 
 	"github.com/dhis2-sre/im-manager/pkg/model"
 )
@@ -45,15 +41,4 @@ func CreateUser(email, password string, userService userServiceUtil, groupServic
 	}
 
 	return nil
-}
-
-func SetCookies(c *gin.Context, tokens *token.Tokens, rememberMe bool, sameSiteMode http.SameSite, hostname string, accessTokenExpirationSeconds int, refreshTokenExpirationSeconds int, refreshTokenRememberMeExpirationSeconds int) {
-	c.SetSameSite(sameSiteMode)
-	c.SetCookie("accessToken", tokens.AccessToken, accessTokenExpirationSeconds, "/", hostname, true, true)
-	if rememberMe {
-		c.SetCookie("refreshToken", tokens.RefreshToken, refreshTokenRememberMeExpirationSeconds, "/refresh", hostname, true, true)
-		c.SetCookie("rememberMe", "true", refreshTokenRememberMeExpirationSeconds, "/refresh", hostname, true, true)
-	} else {
-		c.SetCookie("refreshToken", tokens.RefreshToken, refreshTokenExpirationSeconds, "/refresh", hostname, true, true)
-	}
 }
